+++ Controllers/Website/Account/Logout.php
@@ -1,8 +1,10 @@
 <?php
 
 
-namespace VoicesOfWynn\Controllers;
+namespace VoicesOfWynn\Controllers\Website\Account;
 
+
+use VoicesOfWynn\Controllers\Controller;
 
 class Logout extends Controller
 {
@@ -10,11 +12,12 @@ class Logout extends Controller
     /**
      * @inheritDoc
      */
-    public function process(array $args): bool
+    public function process(array $args): int
     {
         if (isset($_SESSION['user'])) {
             $_SESSION['user']->logout();
         }
         header('Location: /');
+        exit();
     }
 }
