+++ Controllers/Website/Credits.php
@@ -1,18 +1,19 @@
 <?php
 
 
-namespace VoicesOfWynn\Controllers;
+namespace VoicesOfWynn\Controllers\Website;
 
 
+use VoicesOfWynn\Controllers\Controller;
 use VoicesOfWynn\Models\ContentManager;
 
-class Credits extends Controller
+class Credits extends WebpageController
 {
     
     /**
      * @inheritDoc
      */
-    public function process(array $args): bool
+    public function process(array $args): int
     {
         self::$data['base_title'] = 'Credits';
         self::$data['base_description'] = 'Many people worked on this mod, especially many voice actors, both amateur and professional, were needed to complete it. Let\'s appreciate their efforts by looking at their contributions.';
