+++ Controllers/Website/Index.php
@@ -1,13 +1,15 @@
 <?php
 
 
-namespace VoicesOfWynn\Controllers;
+namespace VoicesOfWynn\Controllers\Website;
 
 
-class Index extends Controller
+use VoicesOfWynn\Controllers\Controller;
+
+class Index extends WebpageController
 {
     
-    public function process(array $args): bool
+    public function process(array $args): int
     {
         self::$data['base_title'] = 'Voices of Wynn';
         self::$data['base_description'] = 'Welcome to the webpage of Voices of Wynn - a mod for MMORPG Minecraft server Wynncraft that adds voices to many in-game NPCs.';
