+++ Views/error404.phtml
@@ -26,7 +26,7 @@
     <h1 class="error-header">404 – Not Found</h1>
     <p><i>I don't know where the thing you're looking for is, but I'm sure where it isn't: here.</i></p>
 	<p>Are you sure you typed the URL address correctly?</p>
-    <img src="images/lucio_where_the_frick_are_we.png" alt="An ilustration of two Wynncraft characters looking for something." id="error-picture" /><br>
+    <img src="images/lucio_where_the_frick_are_we.png" alt="An illustration of two Wynncraft characters looking for something." id="error-picture" /><br>
     <span class="art-credit">Art by <a href="https://twitter.com/gleeokenspiel" target="blank_">Gleeokenspiel</a></span><br>
     <a href="https://voicesofwynn.com"><button class="return-button">Back to Homepage</button></a>
 </body>
