+++ routes.ini
@@ -4,38 +4,38 @@
 ; In case a request for a route that is not specified here is made, the 404 error webpage is returned
 
 ; Website routes
-/=Index
-/credits=Credits
-/contents=Contents
-/contents/npc/<0>=Npc?<0>?true
-/contents/npc/<0>/upvote/<1>=Rating?<1>?+
-/contents/npc/<0>/downvote/<1>=Rating?<1>?-
-/contents/npc/<0>/comments/<1>=Comments?<1>
-/contents/npc/<0>/comments/<1>/new=Rating?<1>?c
-/contents/npc/<0>/comments/<1>/delete/<2>=Comments?<2>
-/cast/<0>=Cast?<0>
-/downloads=Downloads
-/faq=Faq
-/login=Login
-/account=Account
-/administration=Administration
-/administration/accounts=Administration?accounts
-/administration/accounts/<0>/clear-bio=Administration?accounts?clear-bio?<0>
-/administration/accounts/<0>/clear-avatar=Administration?accounts?clear-avatar?<0>
-/administration/accounts/<0>/delete=Administration?accounts?delete?<0>
-/administration/accounts/<0>/update-roles/grant/<1>=Administration?accounts?grant-role?<0>?<1>
-/administration/accounts/<0>/update-roles/revoke/<1>=Administration?accounts?revoke-role?<0>?<1>
-/administration/new-account=Administration?new-account
-/administration/npcs=Administration?npcs
-/administration/npcs/manage/<0>=Administration?npc?<0>?false
-/administration/npcs/manage/<0>/delete/<1>=Administration?npc?<0>?false?<1>
-/administration/npcs/manage/<0>/recast/<1>=Administration?npc?<0>?false?<1>
-/logout=Logout
+/=Website\Index
+/credits=Website\Credits
+/contents=Website\Contents
+/contents/npc/<0>=Website\Npc?<0>?true
+/contents/npc/<0>/upvote/<1>=Website\Rating?<1>?+
+/contents/npc/<0>/downvote/<1>=Website\Rating?<1>?-
+/contents/npc/<0>/comments/<1>=Website\Comments?<1>
+/contents/npc/<0>/comments/<1>/new=Website\Rating?<1>?c
+/contents/npc/<0>/comments/<1>/delete/<2>=Website\Comments?<2>
+/cast/<0>=Website\Cast?<0>
+/downloads=Website\Downloads
+/faq=Website\Faq
+/login=Website\Account\Login
+/account=Website\Account\Account
+/administration=Website\Administration\Administration?accounts
+/administration/accounts=Website\Administration?accounts
+/administration/accounts/<0>/clear-bio=Website\Administration\Administration?accounts?clear-bio?<0>
+/administration/accounts/<0>/clear-avatar=Website\Administration\Administration?accounts?clear-avatar?<0>
+/administration/accounts/<0>/delete=Website\Administration\Administration?accounts?delete?<0>
+/administration/accounts/<0>/update-roles/grant/<1>=Website\Administration\Administration?accounts?grant-role?<0>?<1>
+/administration/accounts/<0>/update-roles/revoke/<1>=Website\Administration\Administration?accounts?revoke-role?<0>?<1>
+/administration/new-account=Website\Administration\Administration?new-account
+/administration/npcs=Website\Administration\Administration?npcs
+/administration/npcs/manage/<0>=Website\Administration\Administration?npc?<0>?false
+/administration/npcs/manage/<0>/delete/<1>=Website\Administration\Administration?npc?<0>?false?<1>
+/administration/npcs/manage/<0>/recast/<1>=Website\Administration\Administration?npc?<0>?false?<1>
+/logout=Website\Account\Logout
 
 ; API routes
-/api/unvoiced-line-report/new=Api?newUnvoicedLineReport
-/api/unvoiced-line-report/index=Api?listUnvoicedLineReport
-/api/unvoiced-line-report/raw=Api?getRaw
-/api/unvoiced-line-report/resolve=Api?updateReportStatus
-/api/unvoiced-line-report/reset=Api?resetForwarded
-/api/unvoiced-line-report/accepted=Api?getAcceptedReports
+/api/unvoiced-line-report/new=Api\Api?newUnvoicedLineReport
+/api/unvoiced-line-report/index=Api\Api?listUnvoicedLineReport
+/api/unvoiced-line-report/raw=Api\Api?getRaw
+/api/unvoiced-line-report/resolve=Api\Api?updateReportStatus
+/api/unvoiced-line-report/reset=Api\Api?resetForwarded
+/api/unvoiced-line-report/accepted=Api\Api?getAcceptedReports
